+++ programs/drift/src/state/user.rs
@@ -102,7 +102,15 @@ pub struct User {
     /// User is idle if they haven't interacted with the protocol in 1 week and they have no orders, perp positions or borrows
     /// Off-chain keeper bots can ignore users that are idle
     pub idle: bool,
-    pub padding: [u8; 25],
+    /// number of open orders
+    pub open_orders: u8,
+    /// Whether or not user has open order
+    pub has_open_order: bool,
+    /// number of open orders with auction
+    pub open_auctions: u8,
+    /// Whether or not user has open order with auction
+    pub has_open_auction: bool,
+    pub padding: [u8; 21],
 }
 
 impl User {
@@ -323,6 +331,24 @@ impl User {
             self.last_active_slot = slot;
         }
         self.idle = false;
+    }
+
+    pub fn increment_open_orders(&mut self, is_auction: bool) {
+        self.open_orders = self.open_orders.saturating_add(1);
+        self.has_open_order = self.open_orders > 0;
+        if is_auction {
+            self.open_auctions = self.open_auctions.saturating_add(1);
+            self.has_open_auction = self.open_auctions > 0;
+        }
+    }
+
+    pub fn decrement_open_orders(&mut self, is_auction: bool) {
+        self.open_orders = self.open_orders.saturating_sub(1);
+        self.has_open_order = self.open_orders > 0;
+        if is_auction {
+            self.open_auctions = self.open_auctions.saturating_sub(1);
+            self.has_open_auction = self.open_auctions > 0;
+        }
     }
 }