+++ sdk/src/orderSubscriber/types.ts
@@ -1,0 +1,14 @@
+import { DriftClient } from '../driftClient';
+
+export type OrderSubscriberConfig = {
+	driftClient: DriftClient;
+	subscriptionConfig:
+		| {
+				type: 'polling';
+				frequency: number;
+		  }
+		| {
+				type: 'websocket';
+				skipInitialLoad?: boolean;
+		  };
+};